+++ tests/integration/test_integration.py
@@ -9,6 +9,7 @@ from bs4 import BeautifulSoup
 import dash_dangerously_set_inner_html
 import dash_flow_example
 
+import dash
 from dash import Dash, html, dcc, Input, Output
 from dash.exceptions import PreventUpdate
 
@@ -122,12 +123,12 @@ def test_inin007_meta_tags(dash_duo):
 
     meta = dash_duo.find_elements("meta")
 
-    # -2 for the meta charset and http-equiv.
-    assert len(meta) == len(metas) + 2, "Should have 2 extra meta tags"
+    # -3 for the meta charset, http-equiv and viewport.
+    assert len(meta) == len(metas) + 3, "Should have 3 extra meta tags"
 
-    for i in range(2, len(meta)):
+    for i in range(3, len(meta)):
         meta_tag = meta[i]
-        meta_info = metas[i - 2]
+        meta_info = metas[i - 3]
         assert meta_tag.get_attribute("name") == meta_info["name"]
         assert meta_tag.get_attribute("content") == meta_info["content"]
 
@@ -344,3 +345,54 @@ def test_inin026_graphs_in_tabs_do_not_share_state(dash_duo):
     dash_duo.find_element("#graph2:not(.dash-graph--pending)").click()
 
     until(lambda: '"label": 3' in dash_duo.find_element("#graph2_info").text, timeout=3)
+
+
+def test_inin027_multi_page_without_pages_folder(dash_duo):
+    app = Dash(__name__, pages_folder="")
+
+    # test for storing arbitrary keyword arguments: An `id` prop is defined for every page
+    # test for defining multiple pages within a single file: layout is passed directly to `register_page`
+    # in the following two modules:
+    dash.register_page(
+        "multi_layout1",
+        layout=html.Div("text for multi_layout1", id="text_multi_layout1"),
+        path="/",
+        title="Supplied Title",
+        description="This is the supplied description",
+        name="Supplied name",
+        image="birds.jpeg",
+        id="multi_layout1",
+    )
+    dash.register_page(
+        "multi_layout2",
+        layout=html.Div("text for multi_layout2", id="text_multi_layout2"),
+        path="/layout2",
+        id="multi_layout2",
+    )
+
+    app.layout = html.Div(
+        [
+            html.Div(
+                [
+                    html.Div(
+                        dcc.Link(
+                            f"{page['name']} - {page['path']}",
+                            id=page["id"],
+                            href=page["path"],
+                        )
+                    )
+                    for page in dash.page_registry.values()
+                ]
+            ),
+            dash.page_container,
+        ]
+    )
+
+    dash_duo.start_server(app)
+    # test layout and title for each page in `page_registry` with link navigation
+    for page in dash.page_registry.values():
+        dash_duo.find_element("#" + page["id"]).click()
+        dash_duo.wait_for_text_to_equal("#text_" + page["id"], "text for " + page["id"])
+        assert dash_duo.driver.title == page["title"], "check that page title updates"
+
+    assert not dash_duo.get_logs()